# Conflicts:
#	GIS.cpp
#	GIS.h
#	entities/EntityJsonParser.cpp
#	entities/geometry/Geometry.h
#	tests/CMakeLists.txt
#	tests/GeometryTest.cpp

+++ bonus.error_handling/error_handling_more.cpp
@@ -25,6 +25,6 @@ TEST(ErrorHandling, MixedIdTest) {
 
     auto entityIds = gis.loadMapFile("mixed-id-map.json");
 
-    ASSERT_EQ(entityIds.size(), 2);
+    ASSERT_EQ((int)entityIds.size(), 2);
     ASSERT_TRUE(logContainsError("Couldn't load invalid entity - JSON entity doesn't contain id"));
 }
